requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
azure-cognitiveservices-speech==1.43.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
duckduckgo-mcp-server==0.1.1
edge-tts @ git+https://github.com/Ra1nek/edge-tts_RU.git@31a66b7cf54d4bb724fb65516b88ed1e0c14114b
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
identify==2.6.9
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langdetect==1.0.9
letta-client==0.1.101
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.3
mcp-server-time==2025.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.21.0
openai==1.72.0
openai-harmony==0.0.2
packaging==24.2
platformdirs==4.3.7
pre_commit==4.2.0
propcache==0.3.1
protobuf==6.30.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pysbd==0.3.4
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.26.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.4
scipy==1.15.2
setuptools==80.9.0
shellingham==1.5.4
sherpa-onnx==1.12.6
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
srt==3.5.3
sse-starlette==3.0.2
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
triton==3.2.0
twitchAPI==4.2.0
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uv==0.8.4
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
yarl==1.19.0
chromadb==0.5.23
sentence-transformers==3.0.1
slowapi==0.1.9
